+++ Casks/integra.rb
@@ -1,6 +1,6 @@
 cask "integra" do
-  version "0.8.6"
-  sha256 "618956f6d44102f7625edaf99bce4b58ed9278719488f38f1bb90a4b2e01da4a"
+  version "0.9.0"
+  sha256 "d8533136a646e964e84e8d8fb4bb641dffb89b03c07180b9a7b99d3a61652ef2"
 
   url "https://github.com/Octadira/integra/releases/download/v#{version}/Integra-v#{version}.dmg"
   name "Integra"